STORY-7279: moved function into own file and different place
